fix assertion, start addressing #1087 by using size_t

+++ src/smt/theory_str.cpp
@@ -9194,7 +9194,7 @@ namespace smt {
         // ----------------------------------------------------------------------------------------
         int len = atoi(lenStr.encode().c_str());
         bool coverAll = false;
-        vector<int_vector, true, long long> options;
+        vector<int_vector, true, size_t> options;
         int_vector base;
 
         TRACE("str", tout
